+++ test/exit.ts
@@ -1,0 +1,6 @@
+import 'source-map-support/register'
+import { test } from '../index'
+
+test('exit in test', () => {
+    process.exit(1)
+})